requirements.txt
pytest==3.7.2
pytest-selenium==1.13.0
webdriver-manager==2.3.0
allure-pytest==2.8.6
requests==2.20.0